+++ lib/hello_world.rb
@@ -1,3 +1,7 @@
 def hello(name)
   puts "Hello, #{name}!"
 end
+
+def unused_method
+  puts "I am unused"
+end